Better event handling
Currentlt all events that OpenLayers triggers from
OpenLayers.Map and OpenLayers.Layer are either wrapped
or removed with a good reason. See the inline comments
for more information.

The events from OpenLayers.Layer.Vector aren't wrapped yet.

All events that are triggered on a layer, also trigger those
events on the map, but with the layer that triggered it included.